refactor(capabilities): move inline cap configs into config.rs (#2236)
Extract the five inline-in-mod.rs capability config structs
(anthropic, lifecycle, component, render, rpc/server) each into a
dedicated `<cap>/config.rs`, preserving all cfg-gating, imports,
doc comments, and `pub use` re-exports unchanged. Follows the
established `audio/config.rs` exemplar (ADR-0090).

Closes #2236
